lambda/parser/requirements.txt
boto3
requests
pymongo>=4.0,<5.0
pystache
google-genai
